backend/gateway/lab_service/requirements.txt
Flask
flask-cors
flask_sqlalchemy
python-dotenv
PyJWT